Rename ovsp4rt.h to ovsp4rt_private.h
