+++ ovs-p4rt/sidecar/CMakeLists.txt
@@ -6,12 +6,12 @@
 
 add_library(ovs_sidecar_o OBJECT
     ovsp4rt.cc
-    ovsp4rt.h
-    ovsp4rt_session.cc
-    ovsp4rt_session.h
     ovsp4rt_credentials.cc
     ovsp4rt_credentials.h
+    ovsp4rt_private.h
     ovsp4rt_public.cc
+    ovsp4rt_session.cc
+    ovsp4rt_session.h
     ovsp4rt_utils.cc
     ovsp4rt_utils.h
 )
