+++ ovs-p4rt/sidecar/es2k/ovsp4rt_es2k_public.cc
@@ -1,14 +1,14 @@
 // Copyright 2022-2024 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
-// ES2K-specific implementations of functions declared in ovs-p4rt.h.
+// ES2K-specific functions declared in ovs-p4rt.h.
 //
 
 #include "absl/flags/flag.h"
 #include "openvswitch/ovs-p4rt.h"
-#include "ovsp4rt.h"
 #include "ovsp4rt_credentials.h"
 #include "ovsp4rt_es2k_private.h"
+#include "ovsp4rt_private.h"
 #include "ovsp4rt_session.h"
 #include "ovsp4rt_utils.h"
 #include "p4_name_mapping.h"
